About page and astronaut svg upload. Delete button bug fix
The astronaut svg file is now set up and no longer has the white borders around it. The About page now just has the about svg with all the haikus and a footer. I also made sure that delete buttons arent added to the rows in the 'Other projects that you're involved with' table